beardog_traits::core

Trait SerializableEntity

pub trait SerializableEntity {
    // Required methods
    fn serialize_bytes(&self) -> BearDogResult<Cow<'_, [u8]>>;
    fn deserialize_bytes(bytes: &[u8]) -> BearDogResult<Self>
       where Self: Sized;
    fn format() -> SerializationFormat;
}

Serialization trait with zero-copy optimization

Required Methods

fn serialize_bytes(&self) -> BearDogResult<Cow<'_, [u8]>>

Serialize to bytes with zero-copy when possible

Errors

Returns an error if serialization fails due to invalid data or encoding issues

fn deserialize_bytes(bytes: &[u8]) -> BearDogResult<Self>
where Self: Sized,

Deserialize from bytes

Errors

Returns an error if deserialization fails due to invalid data format or corruption

fn format() -> SerializationFormat

Get serialization format

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.