beardog_traits::security

Trait PolicyEnforcer

pub trait PolicyEnforcer<P> {
    // Required methods
    fn evaluate(
        &self,
        policy: &P,
        context: &PolicyContext,
    ) -> BearDogResult<PolicyDecision>;
    fn validate_policy(&self, policy: &P) -> BearDogResult<()>;
    fn get_policy_metadata(&self, policy: &P) -> PolicyMetadata;
}

Security policy enforcement trait

Required Methods

fn evaluate( &self, policy: &P, context: &PolicyContext, ) -> BearDogResult<PolicyDecision>

Evaluate a policy against a context

Errors

Returns an error if policy evaluation fails due to invalid policy or context

fn validate_policy(&self, policy: &P) -> BearDogResult<()>

Validate policy syntax

Errors

Returns an error if the policy syntax is invalid or malformed

fn get_policy_metadata(&self, policy: &P) -> PolicyMetadata

Get policy metadata