beardog_traits::security

Trait ZeroCopySecure

pub trait ZeroCopySecure {
    // Required methods
    fn verify_in_place(&self, data: &[u8]) -> BearDogResult<bool>;
    fn secure_view<'data>(
        &self,
        data: &'data [u8],
    ) -> BearDogResult<Cow<'data, [u8]>>;
    fn sanitize_in_place(&self, data: &mut [u8]) -> BearDogResult<()>;
}

Zero-copy security operations for performance

Required Methods

fn verify_in_place(&self, data: &[u8]) -> BearDogResult<bool>

Verify data without copying

Errors

Returns an error if verification fails due to invalid data or cryptographic failure

fn secure_view<'data>( &self, data: &'data [u8], ) -> BearDogResult<Cow<'data, [u8]>>

Get secure view of data

Errors

Returns an error if secure view creation fails due to access control violations

fn sanitize_in_place(&self, data: &mut [u8]) -> BearDogResult<()>

Sanitize data in place

Errors

Returns an error if sanitization fails due to unsupported data format